prototype/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
sqlglot==25.24.5
httpx==0.27.2
PyYAML==6.0.2
python-jose[cryptography]==3.3.0
structlog==24.4.0
prometheus-client==0.21.0
opentelemetry-api==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-instrumentation-fastapi==0.48b0
opentelemetry-exporter-otlp-proto-grpc==1.27.0
pytest==8.3.3
pytest-asyncio==0.24.0
